Converted from askama to tera
Tera allows me to parse templates without having to define a struct for each one. Also, it allows hot reloading (which I will eventually implement). Although it is magnitudes slower than askama, it still takes a negligible amount of time, which is a small price to pay for they hugely improved flexibility.